+++ plugins/orchestrator/src/components/InfoDialog.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Button from '@mui/material/Button';
+import { Button } from '@material-ui/core';
 import { useArgs } from '@storybook/preview-api';
 import { Meta, StoryObj } from '@storybook/react';