+++ plugins/orchestrator/CHANGELOG.md
@@ -1,5 +1,19 @@
 ### Dependencies
 
+* **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.15.0
+* **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.0.2
+
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.13.1
+
+### Dependencies
+
+* **@janus-idp/backstage-plugin-orchestrator-form-api:** upgraded to 1.0.1
+* **@janus-idp/backstage-plugin-orchestrator-form-react:** upgraded to 1.0.1
+
+### Dependencies
+
 * **@janus-idp/backstage-plugin-orchestrator-common:** upgraded to 1.14.0
 
 ## @janus-idp/backstage-plugin-orchestrator [1.18.1](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-orchestrator@1.18.0...@janus-idp/backstage-plugin-orchestrator@1.18.1) (2024-08-02)