+++ plugins/rbac/CHANGELOG.md
@@ -1,3 +1,7 @@
+### Dependencies
+
+* **@janus-idp/cli:** upgraded to 1.13.1
+
 ## @janus-idp/backstage-plugin-rbac [1.27.1](https://github.com/janus-idp/backstage-plugins/compare/@janus-idp/backstage-plugin-rbac@1.27.0...@janus-idp/backstage-plugin-rbac@1.27.1) (2024-08-06)